fixed multiple uploader, using swfupload now
Removed old flash multiple uploader in favor of swfupload, which seems more developed and better supported.
If you've over-ridden the default /photos/new.html.haml, you'll need to make some modifications in your version to keep up to date. See CE's version for an example of what to include in that view.